remake: Add vue-router
