+++ src/components/BaseButton.vue
@@ -1,26 +1,23 @@
 <template>
-  <div
-    @click="play"
-    :class="{ playingBtn: playLayer > 0 }"
-  >
+  <div @click="play" :class="{ playingBtn: playLayer > 0 }">
     {{ localizedName }}
   </div>
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from 'vue-property-decorator';
-import { Sound } from '../types';
+import { Component, Prop, Vue } from "vue-property-decorator";
+import { Sound } from "../types";
 
 @Component
 export default class BaseButton extends Vue {
   @Prop() private item!: Sound;
-  private playLayer: number = 0;
-  private lang: string = 'zh';
+  private playLayer = 0;
+  private lang = "zh";
   private mounted() {
     this.lang = this.$i18n.locale;
   }
   get localizedName() {
-    return this?.item.name_l10n![this.lang] || this?.item.name || '';
+    return this?.item.name_l10n![this.lang] || this?.item.name || "";
   }
   private play() {
     if (!this.item) {
@@ -30,20 +27,22 @@ export default class BaseButton extends Vue {
       return;
     }
     let audioFilename;
-    if (typeof (this.item.file) === 'string') {
+    if (typeof this.item.file === "string") {
       audioFilename = this.item.file;
     } else {
-      audioFilename = this.item.file[Math.floor(Math.random() * this.item.file.length)];
+      audioFilename = this.item.file[
+        Math.floor(Math.random() * this.item.file.length)
+      ];
     }
     const audio = new Audio(`assets/${audioFilename}`);
     const that = this;
-    audio.addEventListener('play', () => {
+    audio.addEventListener("play", () => {
       that.playLayer += 1;
-      that.$store.commit('playOne');
+      that.$store.commit("playOne");
     });
-    audio.addEventListener('ended', () => {
+    audio.addEventListener("ended", () => {
       that.playLayer -= 1;
-      that.$store.commit('stopOne');
+      that.$store.commit("stopOne");
     });
     audio.play();
   }
