+++ src/lang/en.ts
@@ -1,6 +1,6 @@
 export default {
-    'Do Not Click Me': 'Do Not Click Me',
-    'GitHub': 'GitHub',
-    'Telegram': 'Telegram',
-    'Toggle chorus mode': 'Toggle chorus mode',
+  "Do Not Click Me": "Do Not Click Me",
+  GitHub: "GitHub",
+  Telegram: "Telegram",
+  "Toggle chorus mode": "Toggle chorus mode"
 };
