+++ src/lang/ja.ts
@@ -1,6 +1,6 @@
 export default {
-    'Do Not Click Me': 'クリックしないでください',
-    'GitHub': 'GitHub',
-    'Telegram': 'Telegram',
-    'Toggle chorus mode': 'Toggle chorus mode',
+  "Do Not Click Me": "クリックしないでください",
+  GitHub: "GitHub",
+  Telegram: "Telegram",
+  "Toggle chorus mode": "Toggle chorus mode"
 };
