+++ src/lang/zh.ts
@@ -1,6 +1,6 @@
 export default {
-  'Do Not Click Me': '没事不要点',
-  'GitHub': '代码猫',
-  'Telegram': '小飞机',
-  'Toggle chorus mode': '多重吟唱模式开关',
+  "Do Not Click Me": "没事不要点",
+  GitHub: "代码猫",
+  Telegram: "小飞机",
+  "Toggle chorus mode": "多重吟唱模式开关"
 };
