+++ src/main.ts
@@ -1,12 +1,14 @@
-import Vue from 'vue';
-import App from './App.vue';
+import Vue from "vue";
+import App from "./App.vue";
+import router from "./router";
 import store from './store';
 import { i18n } from './i18n-setup';
 
 Vue.config.productionTip = false;
 
 new Vue({
+  router,
   store,
   i18n,
-  render: (h) => h(App),
-}).$mount('#app');
+  render: h => h(App)
+}).$mount("#app");
