+++ src/types.ts
@@ -1,6 +1,6 @@
 export interface Sound {
-    name: string;
-    name_l10n?: { [key: string]: string; };
-    file: string | string[];
-    type: string;
+  name: string;
+  name_l10n?: { [key: string]: string };
+  file: string | string[];
+  type: string;
 }
