+++ vue.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  chainWebpack: (config) => {
+    config.plugin("html").tap((args) => {
+      args[0].title = "Starbuttons";
+      return args;
+    });
+  },
+};
